Rename the BC7CompressionMode Quick and Max values
Renamed the BC7CompressionMode Quick and Max values to Fast and Slow.
This makes it clearer to users that the BC7 compression mode is related
the compression speed, and not the amount of image data that it will
discard during compression.

The BC7 compression mode controls the number of BC6H or BC7 modes
DirectXTex will try to use when compressing each 4x4 pixel block.
Using more modes can result in better image quality, at the expense of
taking longer to compress the image data.