Add support for loading and saving cube maps
A cube map is represented as a horizontal or vertical "crossed" image
that contains the cube map faces in a specific order.

The "crossed" image is required to have a 4:3 aspect ratio for
horizontal cube maps or a 3:4 aspect ratio for vertical cube maps, with
the cube map images being square.
For example, a horizontal "crossed" image with 256 x 256 pixel cube map
faces would have a width of 1024 and a height of 768.

Because Paint.NET does not currently allow FileType plugins to disable
UI based on the document size, the cube map setting will be silently
ignored for images that do not meet the size requirements.