Add cancellation support for the progress callback
Changed the DirectXTex progress callback to return a Boolean.
This allows Paint.NET to signal that encoding should be canceled.

Updated the various conversion and encoding methods to abort when the
progress callback returns false.
The multi-threaded CPU encoder will take longer to stop because
OpenMP 2.0 does not support breaking out of a `parallel for` loop.