Reduce the chattiness of the progress reporting
Updated the parallel compression method to only report progress when a
new row is reached.

Previously it was reporting progress when each 4x4 pixel block was
compressed.
This caused Paint.NET to freeze when saving larger images due to the
massive number of progress reporting messages that were being posted
to the UI thread.