Attempt to detect files with the wrong extension
It is not uncommon to see BMP or PNG files with a .dds file extension.

DirectXTex has been modified to return a specific error code for files
that do not have a DDS file signature.
When this error occurs the plugin will attempt to detect the actual
file format and load it using the appropriate PDN FileType handler.