Move the load post-processing into managed code
This reduces memory usage by removing the need for the native code to
copy the output image data or flatten the cube map images.
Undoing the premultiplied alpha is also more efficient as it is only
applied to the images that we read instead of every image in the file.

Changed the output format back to RGBA, Paint.NET will perform the RGBA
to BGRA conversion when it loads the image.