Fix the DdsSaveInfo P/Invoke marshaling on ARM64
The NativeDdsSaveInfo structure must be passed as a ref parameter in
order for it to be marshaled as a pointer to the structure.

The generated marshaling code was passing the native structure by
value instead of as a pointer to the structure, this somehow worked on
x64 but it crashes on ARM64.