Add support for saving the ATI1, ATI2 and RXGB formats
ATI1 is a legacy DX9 four-character-code for the BC4 unsigned format.
ATI2 is a legacy DX9 four-character-code for the BC5 unsigned format.

RXGB is a BC3/DXT5 variant with swizzled image channels, it is used by
some older games (e.g. Doom 3) for storing normal map data.
The red channel data stored in the compressed image's alpha channel and
the compressed image's red channel is set to 0.